rework matched/mismatched models
