+++ caribou_hi/__init__.py
@@ -2,13 +2,15 @@ __all__ = [
     "EmissionModel",
     "AbsorptionModel",
     "EmissionAbsorptionModel",
-    "EmissionAbsorptionFFModel",
+    "EmissionAbsorptionMatchedModel",
+    "EmissionAbsorptionMismatchedModel",
 ]
 
 from caribou_hi.emission_model import EmissionModel
 from caribou_hi.absorption_model import AbsorptionModel
 from caribou_hi.emission_absorption_model import EmissionAbsorptionModel
-from caribou_hi.emission_absorption_ff_model import EmissionAbsorptionFFModel
+from caribou_hi.emission_absorption_matched_model import EmissionAbsorptionMatchedModel
+from caribou_hi.emission_absorption_mismatched_model import EmissionAbsorptionMismatchedModel
 
 from . import _version
